Updated AddSuffix in Secret.yaml
Previous code example for AddSuffix is incorrect and the description is misleading. googleMapsApiKey:AddSuffix(url) would add the URL behind the secret which is not usable in making API calls. AddSuffix is for appending strings to the end of a secret, which is particularly useful for adding extra query parameters to the end of a url after an API key. 
I would have liked to keep out the AddPrefix calls from the updated example but unless you set the entire first part of the URL as the secret this will not be usable. If there is another example where it would fit better to add a prefix to the end of a Secret without directly relating to the URL that would be better example wise, but I can't think of any atm.